- replaced VisualizationReporter with local class Visualizer::Reporter. The API changes also: the new Reporter expects to be given a Visualizer when constructed, rather than a System. This cleans up the very common case where you have to call a few methods of the Visualizer prior to starting a simulation. - fixed some non-const methods that should have been const_iterator - fixed a Windows compiler warning for size_t/int in lodepng

+++ Visualization/VisualizationGUI/lodepng.cpp
@@ -1245,10 +1245,10 @@ static unsigned getHash(const unsigned char* data, size_t size, size_t pos)
 
 static unsigned countInitialZeros(const unsigned char* data, size_t size, size_t pos)
 {
-  int maxCount = MAX_SUPPORTED_DEFLATE_LENGTH;
-  if(maxCount > size-pos)
-    maxCount = size-pos;
-  for(int i = 0; i < maxCount; i++)
+  unsigned maxCount = MAX_SUPPORTED_DEFLATE_LENGTH;
+  if(maxCount > (unsigned)(size-pos))
+    maxCount = (unsigned)(size-pos);
+  for(unsigned i = 0; i < maxCount; i++)
     if(data[pos+i] != 0)
       return i;
   return maxCount;
